Artificial intelligence that evaluates job applications
This project is based on a decision tree structure, developed as an artificial intelligence initiative to objectively assess job applications. The model utilizes machine learning techniques, having been trained on an extensive dataset. Artificial intelligence algorithms analyze application data to evaluate the potential job success of candidates, providing decision-makers with objective information during the hiring process. The goal is to enhance the effectiveness of job application evaluations and select the most suitable candidates. The project aims to contribute to decision-making processes in the business world by promoting a data-driven and objective approach.